+++ src/OpenTelemetry.AutoInstrumentation.Native/environment_variables.h
@@ -11,8 +11,8 @@
 namespace trace {
 namespace environment {
 
-// Sets whether debug mode is enabled. Default is false.
-const WSTRING debug_enabled = WStr("OTEL_DOTNET_AUTO_DEBUG");
+// Sets logging level used by autoinstrumentation loggers
+const WSTRING log_level = WStr("OTEL_LOG_LEVEL");
 
 // Sets the paths to integration definition JSON files.
 // Supports multiple values separated with comma, for example: