+++ src/OpenTelemetry.AutoInstrumentation.Native/logger_impl.h
@@ -37,6 +37,12 @@ private:
     LoggerImpl();
     ~LoggerImpl();
 
+    static inline const WSTRING log_level_none  = WStr("none");
+    static inline const WSTRING log_level_error = WStr("error");
+    static inline const WSTRING log_level_warn  = WStr("warn");
+    static inline const WSTRING log_level_info  = WStr("info");
+    static inline const WSTRING log_level_debug = WStr("debug");
+
 public:
     template <typename... Args>
     void Debug(const Args&... args);
@@ -55,16 +61,12 @@ public:
 
     void Flush();
 
-    void EnableDebug();
     bool IsDebugEnabled() const;
 
     static void Shutdown()
     {
         spdlog::shutdown();
     }
-
-private:
-    bool m_debug_logging_enabled;
 };
 
 template <typename TLoggerPolicy>
@@ -90,8 +92,6 @@ std::string LoggerImpl<TLoggerPolicy>::GetLogPath(const std::string& file_name_s
 template <typename TLoggerPolicy>
 LoggerImpl<TLoggerPolicy>::LoggerImpl()
 {
-    m_debug_logging_enabled = false;
-
     spdlog::set_error_handler([](const std::string& msg) {
         // By writing into the stderr was changing the behavior in a CI scenario.
         // There's not a good way to report errors when trying to create the log file.
@@ -99,14 +99,41 @@ LoggerImpl<TLoggerPolicy>::LoggerImpl()
         // std::cerr << "LoggerImpl Handler: " << msg << std::endl;
     });
 
+    static auto configured_log_level = GetEnvironmentValue(environment::log_level);
+
+    if (configured_log_level == log_level_none)
+    {
+        m_fileout = spdlog::null_logger_mt("LoggerImpl");
+        return;
+    }
+    auto log_level = spdlog::level::info;
+
+    if (configured_log_level == log_level_error)
+    {
+        log_level = spdlog::level::err;
+    }
+    else if (configured_log_level == log_level_warn)
+    {
+        log_level = spdlog::level::warn;
+    }
+    else if (configured_log_level == log_level_info)
+    {
+        log_level = spdlog::level::info;
+    }
+    else if (configured_log_level == log_level_debug)
+    {
+        log_level = spdlog::level::debug;
+    }
+
     spdlog::flush_every(std::chrono::seconds(3));
 
     static auto current_process_name = ToString(GetCurrentProcessName());
-    static auto current_process_id = GetPID();
+    static auto current_process_id   = GetPID();
     static auto current_process_without_extension =
         current_process_name.substr(0, current_process_name.find_last_of("."));
 
-    static auto file_name_suffix = "-" + current_process_without_extension + "-" + std::to_string(current_process_id);
+    static auto file_name_suffix =
+        "-" + current_process_without_extension + "-" + std::to_string(current_process_id);
 
     try
     {
@@ -121,10 +148,11 @@ LoggerImpl<TLoggerPolicy>::LoggerImpl()
         m_fileout = spdlog::null_logger_mt("LoggerImpl");
     }
 
-    m_fileout->set_level(spdlog::level::debug);
+    m_fileout->set_level(log_level);
 
     m_fileout->set_pattern(TLoggerPolicy::pattern, spdlog::pattern_time_type::utc);
 
+    // trigger flush whenever info messages are logged
     m_fileout->flush_on(spdlog::level::info);
 };
 
@@ -161,10 +189,7 @@ template <typename TLoggerPolicy>
 template <typename... Args>
 void LoggerImpl<TLoggerPolicy>::Debug(const Args&... args)
 {
-    if (IsDebugEnabled())
-    {
-        m_fileout->debug(LogToString(args...));
-    }
+    m_fileout->debug(LogToString(args...));
 }
 
 template <typename TLoggerPolicy>
@@ -201,16 +226,11 @@ void LoggerImpl<TLoggerPolicy>::Flush()
     m_fileout->flush();
 }
 
-template <typename TLoggerPolicy>
-void LoggerImpl<TLoggerPolicy>::EnableDebug()
-{
-    m_debug_logging_enabled = true;
-}
 
 template <typename TLoggerPolicy>
 bool LoggerImpl<TLoggerPolicy>::IsDebugEnabled() const
 {
-    return m_debug_logging_enabled;
+    return m_fileout->level() == spdlog::level::debug;
 }
 
 } // namespace shared